Use a recursive mutex for FrameTransport
Summary:
The current implementation of the `FrameTransport` class appears to have race conditions. It seems to be implicated in each of the attached crashes, since a race condition around memory management could result in a lot of related code blowing up at runtime due to UB.

Before this diff, `FramedTransport` synchronizes any time it reads/writes its `std::shared_ptr<FrameProcessor>`, but nothing else is synchronized. This struck me as odd, because it's admitting that `setFrameProcessor()` might be called from different threads than everything else, but most of that method's logic is not in a critical section.

I think we should refactor the class to offer much stronger concurrency guarantees. Ideally this would be with something like `folly::Executor`, but since this class is already mediating two different executor models (`DuplexConnection` vs. `ReactiveSocket`), it doesn't seem beneficial to add a third into the mix. If not that, it'd be great to use a non-recursive mutex here; however, it's difficult to avoid deadlock since a lot of what this class does is invoke ReactiveStreams methods, which may re-enter the `FrameTransport` synchronously.

It is with [great regret](https://fburl.com/recursive-locks) that I would like to propose using a recursive mutex here, **for now**, because it seems strictly better than leaving the critical sections unguarded here and inviting memory corruption.

Reviewed By: alexmalyshev

Differential Revision: D4516181

fbshipit-source-id: 39be7d5e48b80e22963db0f0076890122941a59c