fixing stream ids odd/even
Summary:
Our implementation of odd/even numbers for ReactiveSocket streams is opposite of Java implementation
Not only we need to change it, but we have to make sure that old clients will keep working. To do so we can just disable validation of peer stream ids in registerNewPeerStreamId
I introduced ReactiveSocket version 0.1 to differentiate the old clients which have the odd/even stream numbers wrong and the new clients which have the
stream numbers right so we can run registerNewPeerStreamId no problem.

Reviewed By: yschimke

Differential Revision: D4535375

fbshipit-source-id: e350550e9090698ed1354a1efd28a2b9f6e50813