Removing requestSubscription from ReactiveSocket
Summary:
requestSubscription is replaced with requestStream. Both of the interactions have the same semantic and exchange the same frames, so not only the application code can replace subscription with stream, but also we can make it backward compatible
on protocol level by deserializing SUBSCRIPTION frame as STREAM frame. This deserializing is done in FrameParser_v0.

Depends on D4611318.

This is up for review.
I am not going to land this diff until D4611318 is fully deployed with the gateway.
Also testing on mobile makes sense only after D4611318 is deployed at least to staging tier and mobile is directed to it.

Reviewed By: alexmalyshev

Differential Revision: D4611352

fbshipit-source-id: f5d78633f0b090bc8913df40127a7838cba4b925