Add support for closing a ReactiveSocket with a connection error
Summary:
ReactiveSocket::{disconnect,close}() both assume normal behavior, that no error
has occurred.  Add ReactiveSocket::closeConnectionError(), which sends a
CONNECTION_ERROR error frame and closes the socket.  Need to grab a reference to
`this` in ConnectionAutomaton::connect() as the socket can be closed when
processing a connection frame.  Add a test that would previously fail with a
heap-use-after-free without the extra ref.

Reviewed By: stupaq-fb

Differential Revision: D4614953

fbshipit-source-id: fa9305578fb26b6b9a5b69ba16a4cdde1b309291