Removing frame serialization from stream automata
Summary:
This is a half of the effort to remove dependency on *_Frame types from the stream automata.
The automatons will call virtual methods to write on the wire. Also in the next change they will not parse frames. The framing will not be exposed outside of the ConnectionAutomaton.

Tested with unit tests.
Closes https://github.com/ReactiveSocket/reactivesocket-cpp/pull/327

Differential Revision: D4754884

Pulled By: lehecka

fbshipit-source-id: 97e21583d11a17e53664d06e09885f2935c4a523