wrapping server resume parameters to ResumeParams class
Summary:
I will be adding more parameters in tryResumeServer method and it feels bad that so many values are coppied and plumbed thru. It will be much easier if the "sender" sends them as a package in the new ResumeParameters class.
This is in preparation for the following diff.

Depends on D4728760.

Reviewed By: vitaut

Differential Revision: D4728913

fbshipit-source-id: 1580e617b83d1462b64e382c969d622bb62a0550