protocol version detection for framed duplex connection
Summary:
This is the second half of the autodetection mechanism for the protocol version. FramedDuplexConnection is
writing and reading the frame boundary and as the boundary in v1.0 is encoded with 3 bytes, previously 4 bytes.
Also previously the frame length included the 4 bytes prefix, newly it is excluded.
The autodetection mechanism is not turned on for the gateway with this diff.

Depends on D4735186.

Reviewed By: alexmalyshev

Differential Revision: D4735191

fbshipit-source-id: 4f519b36876078bff95f27f7c3c5f233c1528615