Start making RSocketServer thread-safe
Summary:
Switches out the vector<unique_ptr<ReactiveSocket>> with a
Synchronized<unordered_set<ReactiveSocket*>>.  This lets us erase
ReactiveSockets efficently, and do insertion/deletion in a thread-safe manner.

Also adds RSocketServer::unpark(), to unblock a call of
RSocketServer::startAndPark().  I modified the stream-hello-world example to
start the server on a separate thread, while the main thread waits for a newline
input to unpark the server and stop it.  It works, but TcpConnectionAcceptor is
crashing on shutdown, I'll look into that on a separate diff.

Reviewed By: lehecka

Differential Revision: D4813381

fbshipit-source-id: 19d99555fc7fbee38f21ce283b282764808a0e11