fixing FramedReader
Summary:
I introduced the bug when fixing android build errors. So initially the diff didn't have the bug and all tests were passing.
After fixing the android issue (it didn't like mixing signed and unsigned types) the bug was that I casted "uint8_t << 24" to uint8_t which would obviously always return 0 and not a big number as expected.

Reviewed By: alexmalyshev, somasun

Differential Revision: D4842132

fbshipit-source-id: af67d2f3f0b87577ce6835d14af5c6c74c01f1b1