Allow sending/receiving frames without having a frame processor
Summary:
If we unset the frame processor we can no longer send terminal frames. However if we unset it we risk crashing if we receive any frames while it's unset (this happens outside the stack frame due to a thread hop).

Instead we explicitly support the case of no frame processor set by keeping a queue of incoming frames and flushing those when the processor is back.

Reviewed By: yschimke

Differential Revision: D4921303

fbshipit-source-id: bc3da38242540348318a1d754e64eba8c95639af