Slightly saner {Flowable,Observable}Operator shutdown semantics (#564)
* Slightly saner {Flowable,Observable}Operator shutdown semantics

An operator is between an upstream Flowable and a downstream subscriber.  The
stream of signals it receives from one is not synchronized with the other.  We
can't assume that we have a downstream subscriber when
FlowableOperator::onComplete() is called from the upstream.

We consider an operator "terminated" if it has lost its references to its
upstream and downstream.  A terminated operator can still have methods called
upon it, so we need to make sure we don't send any signals from one side to the
other.

Enables the Observable.SimpleTake test, which previously hit ASAN errors because
of this bad behavior.

* Add TerminateState + terminateImpl()

* Attempt at fixing gcc-4.9 build

* Second attempt at appeasing gcc-4.9