Fix hang when destroying RSocketConnectionManager off of the client IO thread (#675)
* Fix destroying RSocketConnectionManager off of the client IO thread

Or really, any IO thread that's managing one of the state machines inside the
connection manager.

To do so, the thread running ~RSocketConnectionManager needs to close the state
machines using it synchronously.  To do that, it has to relinquish the
connections map lock.  We'll move out all connections from the map first, then
close them all (either sync or async).  Then we'll use the moral equivalent of a
latch (atomic_int + baton) to determine when all of them have actually closed.

* Run onClosed() handlers last

So we don't accidentally destroy the state machine before its streams and frame
transport