Add fire-and-forget benchmark (#742)
Pretty similar to the request-response one, except this one puts the latch on
the server's responder.

Adds `(void)n;` statements to benchmarks so the compiler shuts up about unused
variables.  Note: We can remove the `n` argument entirely from the macro, but
that'll just make the compiler cry about macros being called with the wrong
number of arguments.