Make ~ConnectionSet synchronous
Summary:
It's asynchronous right now, and this ends up making server shutdown a lot
harder, as we can't reason about what the state of connections and subscribers
are after ~RSocketServer.

Changes the std::weak_ptr<ConnectionSet> field in RSocketStateMachine to a
ConnectionSet*.  We make a hard requirement that if an RSocketStateMachine is
registered to a set, then the set must outlive it.  Also saves 8 bytes per
RSocketStateMachine which is nice.

This only affects RSocketServer shutdown, so there shouldn't be any visible
change in perf except for servers with extremely tiny lifetimes.  Each
individual connection is still closed off in parallel, only the thread running
~ConnectionSet will block.

Reviewed By: dymk

Differential Revision: D6688501

fbshipit-source-id: 0f59f5281586c7cc1cfcb331f36094f083497812