removing template param from ObservableOperator
Summary:
ObservableOperator was storing a pointer to the parent Observable which was used only in less than 50% of the cases. By removing the pointer there is no need to pass in the 3rd template argument.
This change also reduces denerated symbols as Observable<U, D> is more generic than Observable<U, D, T>.

Depends on D6821738.

Reviewed By: phoad

Differential Revision: D6822101

fbshipit-source-id: 86578b812f40fac3d800d8ac7d02d7501ecd7259