Merge StreamState into RSocketStateMachine
Summary:
StreamState as an abstraction is an awkward mess.  It holds the map of stream
IDs to stream state machines, but it never uses it.  All it does is expose it as
a public field.  It also holds the queue of pending frames to be output, which
makes for an awkward combination with the stream machine map.

Move the two fields over to RSocketStateMachine instead.  I'm intentially moving
the `StreamId -> StreamStateElem` map right next to the `StreamId ->
StreamFragmentAccumulator` map because I think they should be merged in the
future.

Reviewed By: phoad

Differential Revision: D6956411

fbshipit-source-id: f8f71bf409ac7ab6a9abe7d5f589101501a00184