Pass folly::StringPiece to Payload ctor instead of const std::string&
Summary:
We're copying the string into a folly::IOBuf with folly::IOBuf::copyBuffer().
It's better to take folly::StringPiece so that places where we have a string
literal aren't forced to allocate a std::string object.

Only affects ERROR frames so I'm not expecting any perf improvements, but I
think this is still good to do.

Reviewed By: dymk

Differential Revision: D7034187

fbshipit-source-id: 67545f9d08c5bc8c4d1620e1326c03b19ae06d54