Concat operator should release ownership of the Subscriber
Summary:
In this newly added unit test, even if we cancel the `subscriber` at the end, the subscriber object wasn't get destroyed.

With this update, I delete all cyclic links when either subscription completes or get cancelled. So, the external `subscriber` gets released.

We will use `concat` operator in Thrift.

Reviewed By: alexmalyshev

Differential Revision: D7253716

fbshipit-source-id: 44ecc834b78ce04ddc490729722104e0e5bbbe5d