Don't keep a shared_ptr to RSocketStateMachine in StreamStateMachineBase
Summary:
This cycle isn't necessary.  In almost all cases, the RSocketStateMachine (as a
StreamsWriter) outlives the StreamStateMachineBase.  This means we should be
able to keep a raw back-reference (StreamsWriter&) instead of a full shared_ptr.

There's a few issues with removeFromWriter() being called more than once, and
when the StreamsWriter& back reference is no longer pointing to valid memory.
We can dodge that by setting the stream's ID to 0.  Add a bunch of debug checks
to ensure we're not trying to write stream frames with ID 0.

Shrinks StreamResponder from 72 bytes down to 64 bytes.

Reviewed By: phoad

Differential Revision: D7052606

fbshipit-source-id: 46617fe0e20b4d50017bb61dfd5d634f00c315a3