Subscribing multiple times to the resultant Stream objects of
Summary:
We needed to clone the initial `request`, so each subscribe call to the result can use use it.

Without cloning the `request`, the first subscribe causes the remote call with correct data, but it also destroys the `request` object. So subscribing again causes crash instead of sending a new request.

We need this functionality for `retry` operation in Yarpl.

Reviewed By: alexmalyshev

Differential Revision: D7353581

fbshipit-source-id: 2d53e90dbce3a2d59386645dfde45be7c2b57a5b