Back out "Don't keep a shared_ptr to RSocketStateMachine in StreamStateMachineBase"
Summary:
I'm seeing crashes where stream state machines are accessing the connection
state machine after their stream IDs have been set to zero, and just an increase
of crashes in general.  I think this diff is the culprit.

The diff in question was just a small optimization, so it's fine to back it out.

Reviewed By: phoad

Differential Revision: D7407376

fbshipit-source-id: b9c9b17c430c4d5ac509ce54a06f7223ab12402e