Wipe the StreamsFactory class, eliminates cyclic reference with RSocketStateMachine
Summary:
RSocketStateMachine keeps StreamsFactory as a shared_ptr internally. StreamsFactory also keeps a reference to the RSocketStateMachine from which it calls shared_from_this() and provides to other StreamStateMachines.

We did not need this cyclic reference, so this update gets rid of StreamsFactory class.
I also added unit tests for requestStream and responseStream related functions in RSocketStateMachine.

Reviewed By: alexmalyshev

Differential Revision: D7748850

fbshipit-source-id: d0a154886e02a6feb29879418da17845b4cf2ee7