Throw instead of crash on stream ID overflow
Summary:
We shouldn't be crashing if we happen to overflow on stream IDs.  However we
should crash if we create a new stream ID and it somehow already exists in the
map.

Context - I'm seeing a lot of crashes around creating StreamRequesters and
adding them to the streams map.  It _seems_ like somewhere we're failing to
insert the StreamRequester into the map, and the only way that can happen is if
we have a duplicate ID somewhere...  Hopefully this might give us more
meaningful crashes on mobile.  Otherwise it shouldn't do anything new.

Reviewed By: phoad

Differential Revision: D7785902

fbshipit-source-id: 4077f16b3e0ad8edb02c2a6f3b513b5f129133ab