Delete ProtocolVersion::Current()
Summary:
We had this method to be able to set a default version from the command line.
In reality, we don't want to default to anything, we want to run whatever
version the client tells us to run, and that's either 0.1 or 1.0.

Start with getting rid of this Current() static method, and replacing all uses
of it with ProtocolVersion::Latest, which is what it is in all cases already.

Reviewed By: phoad

Differential Revision: D7797896

fbshipit-source-id: 713a9382ddc1049287023154bb2d2cf6392e7939