Make RSocketStateMachine don't depend on ConnectionSet
Summary:
If we add a generic CloseCallback to RSocketStateMachine and use this callback to inform that the state machine is closed, then we can get rid of the ConnectionSet pointer in the RSocketStateMachine.

Why? I want to use wangle::ConnectionManager to manage the connections as it's already used as the connection manager in Thrift. This hard dependency to ConnectionSet wasn't allowing replacement of the connection manager.

Reviewed By: alexmalyshev

Differential Revision: D7797273

fbshipit-source-id: ea66b840572e66076b05cf7d9f346df84781eb5b