Split ProtocolVersion into its own module in rsocket/framing
Summary:
This belongs in with the rest of the framing code.  Right now it's split between
the mess that is rsocket/internal/Common.h but also
rsocket/framing/FrameSerializer.cpp for some reason.

Getting us closer to having rsocket/framing be its own library, separate from
the rest of the rsocket/ code.

Reviewed By: phoad

Differential Revision: D7802875

fbshipit-source-id: ab50b4eb020da22e2799487a86d8a06e8e7910d2