Get FrameFlags out of Payload
Summary:
Mostly a refactor.  Payload is functionally just a pair of IOBuf pointers.  It
shouldn't know about frame flags, how to produce them, or check them.  Payload
is something we expose to users in the top-level RSocket client APIs, whereas
FrameFlags is not.

Move Payload::getFlags() and Payload::checkFlags() over to Frame.{cpp,h}.  At
the same time, make checkFlags() throw when the METADATA flag doesn't match the
existence of the metadata buffer.  We shouldn't be able to construct a Frame of
this form.

Reviewed By: phoad

Differential Revision: D7942067

fbshipit-source-id: 77d3ab2d160bd5aa0a492622960e2f7f07ee79df