Preserve error's content
Summary:
RSocket was converting the error payload into string and representation of the error to always a runtime_error.
This generalization causes loss of data.

This diff adds rsocket::ErrorWithPayload struct. This struct extends std::exception.
It keeps the Payload, so it is possible to send/receive encoded data through this Payload.

Where will this help?
As the transport doesn't prune the data, the receiver subscriber can assume the received data is as it is sent from the responder. If the responder has encoded the error, the receiver can decode it.

Reviewed By: andriigrynenko

Differential Revision: D8417601

fbshipit-source-id: af413766dbfac676c54b6b4aaf56b459fe2b9c04