Convert SFINAE to static_assert in {Map,Filter,Reduce}Operator class templates
Summary: This replaces some instances of SFINAE with `static_assert` in class template parameters. The result is that invalid template arguments will fail early at compile time with a clear compile error. This is because the compiler will select these classes as a match during template substitution even if the supplied template arguments do not constitute a valid input.

Reviewed By: yfeldblum, phoad

Differential Revision: D8517995

fbshipit-source-id: 0d95ced940c89311ea905f6b6e29ceb9ad2e4ca7