Move Resume{IdentificationToken,Position} and StreamId out of internal/Common.h
Summary:
Trying to get the build rules to make sense, and to stop using the monolithic
internal/Common.h as much as possible.

Moves ResumeIdentificationToken to its own module, StreamId to FrameHeader.h,
and ResumePosition to Frame.h.  Those places made the most sense to me as to how
they're used and how complicated ResumeIdentificationToken was.

Deletes getStreamType() as that wasn't used.

Reviewed By: lexs

Differential Revision: D8812638

fbshipit-source-id: a588a0782608c630f2ca2d1f5cb685d203d8bfa7