Don't send onDisconnect callback before resuming
Summary:
If we have an open connection when a resume comes in we drop the old
connection as expected. However we also send a onDisconnected callback to the
server which makes it think it should pause the session. It pauses everything
only to unpause it moments later as we were in the process of resuming.

Reviewed By: phoad

Differential Revision: D9267026

fbshipit-source-id: 8894291a84da3ab0072ed87c1b783246092fcc17