Reapplying only boiler plate code from D8955216
Summary:
Original commit changeset: 8703161f6d79
The original diff D8955216 landed but caused memory regression and later was reverted in D9262496. In order to find the problematic piece of the original code,
I am splitting it to the boiler plate code and the rest. This diff contains only the boiler plate code change. It is adding a parameter ProtocolVersion to the interface method
ConnectionFactory::connect and all its implementations. No behavior is added or changed.

Reviewed By: phoad

Differential Revision: D9501071

fbshipit-source-id: 172d30fda0c08896d198c75d3f063dbe1e95f613