Reworked and reapplying D8955216
Summary:
The original diff D8955216 caused memory regression in RTGW. In this diff I redesigned the control/child stream handler to not require changes in the Http2StreamHandler (shared class with the Http2ServerStreamHandler, which is currently used in prod).
I also removed the need to use std::deque which was increasing the memory usage. The frame chunk accumulation and buffering is done purely thru IOBufQueue which doesn't require any extra allocations.

The control stream handler implementation is disabled by default and the new code is not touching the http server handlers used in RTGW.

Reviewed By: sarangbh

Differential Revision: D9544659

fbshipit-source-id: 2f225a44d61b3a9753567f1f08734d06018cca41