do not use jemalloc with clang on Travis CI builds
Summary:
The Travis CI builds use Ubuntu trusty, and the version of jemalloc that it
ships with is 4+ years old and does not work with clang+libstdc++.  clang
thinks that jemalloc's definitions of memalign and other functions conflict
with those provided by libstdc++.  This was fixed in upstream jemalloc in
2015 (jemalloc commit ae93d6bf3), but this fix is not available in Ubuntu
trusty.

This hasn't caused problems until now since the autoconf-based build of folly
did not use jemalloc unless arguments were supplied to configure explicitly
asking for jemalloc to be used.  The CMake-based build of folly uses jemalloc
automatically if it is available.

Reviewed By: phoad

Differential Revision: D10408985

fbshipit-source-id: 80afedcc079d1cd638863f6091ba322b708ccd85