travis: build with clang-6.0 instead of clang-4.0
Summary:
folly does not compile with clang versions older than 6.0 any more.

This also changes the travis.yml file to set CC and CXX in the environment
so that folly and gmock get built with the intended compiler instead of being
compiled with the default system compiler (gcc 4.8.4 on Ubuntu trusty).  folly
no longer compiles with gcc 4.8 either.

Reviewed By: phoad

Differential Revision: D10435312

fbshipit-source-id: 11eab6d229668fbdad44e3c6cf7bbb82cf175e8a