Remove HHWheelTimer::Callback::getCurTime and its overrides
Summary:
This function is being overridden in only one class (with a duplicate) that is used in only a single test with no clear semantics of what is being tested (the only effect it has is that cascading logic will see 0, which can be achieved with a simple sleep).

Removing this customization point allows us to avoid doing additional call to steady_clock::now per each timeout (stacked diff) and thus improving performance by ~2x (measured as part of Fibers benchmark after migrating FiberManager to use HHWheelTimer, also stacked).

Reviewed By: jmswen

Differential Revision: D13624362

fbshipit-source-id: ef3abd066d2a4f46b369f65d64dd58399ebf5e46