rsocket: remove -pedantic from the cmake build
Summary:
This was a real PITA to track down.  The problem is this:

* Ubuntu 16 has a version of jemalloc installed that doesn't annotate
  `malloc` with `throw ()`.
* Folly unconditionally pulls in that jemalloc header
* rsocket/yarpl, when compiling with `-pedantic` enabled, generate
  a compiler error about the difference between the `stdlib.h`
  `malloc() throw();` and the jemalloc `malloc();`

I tried to deal with this by probing for the conflict in the folly
cmake rules (D13795199), but cannot do so without turning on `-pedantic`
for folly and potentially propagating that into all downstream consumers of
folly.

`-pedantic` doesn't generally provide a lot of value and is hard
to reason about which individual checks are being enforced, and
importantly, doesn't provide a way to enable/disable those checks
for sections of code that should be excepted from those rules.

So, the easiest thing to do is remove its use from rsocket.

Reviewed By: yfeldblum

Differential Revision: D13800522

fbshipit-source-id: 552334d2b6cc62834c687e91e8f5af890e79608e