Eliminate accessing Connection after the KeepAliveTimer is destroyed
Summary:
KeepAliveTimer causes RSocket to crash as it tries to access Connection_ object, possible after the destruction.
It shows that the `generation_` check did not work for this case.

As seen, there is already a check to eliminate such failure by taking a reference to the connection_, before calling a callback.
So, make it more generalized.

Ref crash stack trace: P61856171

Reviewed By: sarangbh

Differential Revision: D14627149

fbshipit-source-id: 8fb4b1c22212ac468314530c762d9a62e66645e6