Re-re-sync
Note that shipit wants to restore the `/build` line to the `.gitignore`
which we know is bad.  I'm not sure why it wants to do that even after
removing that line from the version of the file in fbsource; perhaps
shipit is looking at an earlier commit?

Let's try this and see if it makes shipit any happier.